GRangesDeferred Class

Enables deferred reading of BAM files by caching after performing file operations.

GRangesDeferred(cache = GRanges(), operation)

has_cache_contents(.data)

get_cache(.data)

get_operation(.data)

Arguments

cache
a GRanges object

operation
an environment containing operations to perform to generate the cache.

.data
a GRangesDeferred object

Details

To see what is loaded in the cache use get_cache and to see the file operations being performed use get_operation. To check whether the cache is full use has_cache_contents. By default the show method for a GRangesDeferred object shows what's in the cache.

See also

read_bam