Expand list-columns in a Ranges object

Expand list-columns in a Ranges object

# S3 method for GenomicRanges
unnest(data, ..., .drop = FALSE, .id = NULL,
  .sep = NULL)

Arguments

data
A Ranges object

...
list-column names to unnest

.drop
Determines whether other list columns will be dropped. By default unnest will keep other list columns even if they are nested.

.id
A character vector of length equal to number of list columns. If supplied will create new column(s) with name .id identifying the index of the list column (default = NULL).

.sep
Combine name of nested Ranges with name of list column seperated by .sep, currently not implemented.

Value

a GRanges object with expanded list columns

Examples

grng <- as_granges(data.frame(seqnames = "chr1", start = 20:23, width = 1000))
grng <- mutate(grng, exon_id = IntegerList(a = 1, b = c(4,5), c = 3, d = c(2,5)))
unnest(grng)

#> GRanges object with 6 ranges and 1 metadata column:
#>       seqnames     ranges strand |   exon_id
#>          <Rle>  <IRanges>  <Rle> | <integer>
#>   [1]     chr1 [20, 1019]      * |         1
#>   [2]     chr1 [21, 1020]      * |         4
#>   [3]     chr1 [21, 1020]      * |         5
#>   [4]     chr1 [22, 1021]      * |         3
#>   [5]     chr1 [23, 1022]      * |         2
#>   [6]     chr1 [23, 1022]      * |         5
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

unnest(grng, .id = "name")

#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames     ranges strand |   exon_id        name
#>          <Rle>  <IRanges>  <Rle> | <integer> <character>
#>   [1]     chr1 [20, 1019]      * |         1           a
#>   [2]     chr1 [21, 1020]      * |         4           b
#>   [3]     chr1 [21, 1020]      * |         5           b
#>   [4]     chr1 [22, 1021]      * |         3           c
#>   [5]     chr1 [23, 1022]      * |         2           d
#>   [6]     chr1 [23, 1022]      * |         5           d
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths