Read a BAM file

Read a BAM file

read_bam(file, index = file, paired = TRUE, flag_reads = bam_flag_reads(),
  select_reads = bam_select(), overlap_ranges = NULL)

Arguments

file
The path to a BAM file.

index
The path to the index file of the BAM. Must be given without the '.bai' extension.

paired
Whether to treat the BAM file as paired end (TRUE) or single end (FALSE.

flag_reads
The reads that will be returned by selecting combinations in bam_flag_reads.

select_reads
The fields and tags in the BAM file that will be returned as metadata columns in the returned GRanges object. Determined by bam_select.

overlap_ranges
Only return records that overlap the given Ranges object. The returning GRanges object will be annotated with a column called which_label to identify the range in overlap_ranges that the read overalps.

Details

By default for paired = TRUE, read_bam will select reads that are mapped, paired and have mapped mates. For paired = FALSE, read_bam will select reads that are mapped. This function is wrapper to the readGAlignment functions in GenomicAlignemnts.

See also

readGAlignments