Set the width of a genomic interval

Set the width of a genomic interval

set_width(x, width)

Arguments

x
a Ranges object, to fix by either the start, end or center of an interval use anchor_start(x), anchor_end(x), anchor_center(x). To leave strand fixed use either anchor_3p(x) or anchor_end(x).

width
the amount to alter the width of a Ranges object by. Either an integer vector of length 1 or an integer vector the same length as x.

See also

resize

Examples

rng <- Ranges(data.frame(start=c(2:-1, 13:15), width=c(0:3, 2:0)))
set_width(rng, width = 10)

#> IRanges object with 7 ranges and 0 metadata columns:
#>           start       end     width
#>       <integer> <integer> <integer>
#>   [1]         2        11        10
#>   [2]         1        10        10
#>   [3]         0         9        10
#>   [4]        -1         8        10
#>   [5]        13        22        10
#>   [6]        14        23        10
#>   [7]        15        24        10

set_width(anchor_start(rng), 10)

#> IRanges object with 7 ranges and 0 metadata columns:
#>           start       end     width
#>       <integer> <integer> <integer>
#>   [1]         2        11        10
#>   [2]         1        10        10
#>   [3]         0         9        10
#>   [4]        -1         8        10
#>   [5]        13        22        10
#>   [6]        14        23        10
#>   [7]        15        24        10

set_width(anchor_end(rng), 10)

#> IRanges object with 7 ranges and 0 metadata columns:
#>           start       end     width
#>       <integer> <integer> <integer>
#>   [1]        -8         1        10
#>   [2]        -8         1        10
#>   [3]        -8         1        10
#>   [4]        -8         1        10
#>   [5]         5        14        10
#>   [6]         5        14        10
#>   [7]         5        14        10

set_width(anchor_center(rng), 10)

#> IRanges object with 7 ranges and 0 metadata columns:
#>           start       end     width
#>       <integer> <integer> <integer>
#>   [1]        -3         6        10
#>   [2]        -4         5        10
#>   [3]        -4         5        10
#>   [4]        -5         4        10
#>   [5]         9        18        10
#>   [6]         9        18        10
#>   [7]        10        19        10

grng <- Ranges(data.frame(seqnames = "chr1",
                         strand = c("+", "-", "-", "+", "+", "-", "+"),
                         start=c(2:-1, 13:15),
                         width=c(0:3, 2:0)))
set_width(anchor_3p(grng), 10)

#> GRanges object with 7 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1  [ 2, 11]      +
#>   [2]     chr1  [ 1,  1]      -
#>   [3]     chr1  [ 0,  1]      -
#>   [4]     chr1  [-1,  8]      +
#>   [5]     chr1  [13, 22]      +
#>   [6]     chr1  [14, 14]      -
#>   [7]     chr1  [15, 24]      +
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

set_width(anchor_5p(grng), 10)

#> GRanges object with 7 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1  [ 2,  1]      +
#>   [2]     chr1  [-8,  1]      -
#>   [3]     chr1  [-8,  1]      -
#>   [4]     chr1  [-1,  1]      +
#>   [5]     chr1  [13, 14]      +
#>   [6]     chr1  [ 5, 14]      -
#>   [7]     chr1  [15, 14]      +
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths