Subset a Ranges object

Unlike dplyr's filter, the filter for a Ranges is only valid for a single logical expression. For any Ranges objects filter can act on all core components of the class including start, end, width (for IRanges) or seqnames and strand (for GRanges) in addition to metadata columns. If the Ranges object is grouped filter will act on each parition of the data.

# S3 method for GenomicRanges
filter(.data, expr)

# S3 method for Ranges
filter(.data, expr)

# S3 method for GRangesGrouped
filter(.data, expr)

# S3 method for IRangesGrouped
filter(.data, expr)

Arguments

.data
A Ranges object

expr
a valid logical expression to act on .data

Value

a Ranges object

See also

filter

Examples

set.seed(100)
df <- data.frame(start = 1:10, width = 5,  seqnames = "seq1",
strand = sample(c("+", "-", "*"), 10, replace = TRUE), gc = runif(10))
rng <- Ranges(df)
rng %>% filter(strand == "+")

#> GRanges object with 4 ranges and 1 metadata column:
#>       seqnames    ranges strand |                gc
#>          <Rle> <IRanges>  <Rle> |         <numeric>
#>   [1]     seq1  [ 1,  5]      + | 0.624996477039531
#>   [2]     seq1  [ 2,  6]      + | 0.882165518123657
#>   [3]     seq1  [ 4,  8]      + | 0.398487901547924
#>   [4]     seq1  [10, 14]      + | 0.690290528349578
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

rng %>% filter(gc > mean(gc))

#> GRanges object with 5 ranges and 1 metadata column:
#>       seqnames    ranges strand |                gc
#>          <Rle> <IRanges>  <Rle> |         <numeric>
#>   [1]     seq1  [ 1,  5]      + | 0.624996477039531
#>   [2]     seq1  [ 2,  6]      + | 0.882165518123657
#>   [3]     seq1  [ 5,  9]      - |  0.76255108229816
#>   [4]     seq1  [ 6, 10]      - | 0.669021712383255
#>   [5]     seq1  [10, 14]      + | 0.690290528349578
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

rng %>% group_by(strand) %>% filter(gc > mean(gc))

#> GRangesGrouped object with 4 ranges and 1 metadata column:
#> Groups: 
#>       seqnames    ranges strand |                gc
#>          <Rle> <IRanges>  <Rle> |         <numeric>
#>   [1]     seq1  [ 2,  6]      + | 0.882165518123657
#>   [2]     seq1  [10, 14]      + | 0.690290528349578
#>   [3]     seq1  [ 5,  9]      - |  0.76255108229816
#>   [4]     seq1  [ 6, 10]      - | 0.669021712383255
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths