Group a Ranges by one or more variables

group_by takes a Ranges object and converts it to a RangesGrouped object. Operations are performed on these objects by their group. ungroup removes grouping.

# S3 method for GRanges
group_by(.data, ...)

# S3 method for GRangesGrouped
ungroup(x, ...)

# S3 method for IRanges
group_by(.data, ...)

# S3 method for IRangesGrouped
ungroup(x, ...)

Arguments

.data
a Ranges object

...
Variable names to group by. These can be either metadata columns or the core variables of a Ranges.

x
a Ranges object

Value

a GroupedRanges object

Examples

df <- data.frame(start = 1:10, width = 5,  seqnames = "seq1",
strand = sample(c("+", "-", "*"), 10, replace = TRUE), gc = runif(10))
rng <- as_granges(df)
rng

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1  [ 1,  5]      - | 0.488305994076654
#>    [2]     seq1  [ 2,  6]      * | 0.928505074931309
#>    [3]     seq1  [ 3,  7]      - | 0.348691981751472
#>    [4]     seq1  [ 4,  8]      * | 0.954157707514241
#>    [5]     seq1  [ 5,  9]      - | 0.695274139055982
#>    [6]     seq1  [ 6, 10]      + | 0.889453538926318
#>    [7]     seq1  [ 7, 11]      * | 0.180407245177776
#>    [8]     seq1  [ 8, 12]      * | 0.629390850430354
#>    [9]     seq1  [ 9, 13]      - | 0.989564136601985
#>   [10]     seq1  [10, 14]      + | 0.130288870073855
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# grouping does not change appearance or shape or Ranges
rng_by_strand <- rng %>% group_by(strand)
rng_by_strand

#> GRanges object with 10 ranges and 1 metadata column:
#> Groups: strand
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1  [ 1,  5]      - | 0.488305994076654
#>    [2]     seq1  [ 2,  6]      * | 0.928505074931309
#>    [3]     seq1  [ 3,  7]      - | 0.348691981751472
#>    [4]     seq1  [ 4,  8]      * | 0.954157707514241
#>    [5]     seq1  [ 5,  9]      - | 0.695274139055982
#>    [6]     seq1  [ 6, 10]      + | 0.889453538926318
#>    [7]     seq1  [ 7, 11]      * | 0.180407245177776
#>    [8]     seq1  [ 8, 12]      * | 0.629390850430354
#>    [9]     seq1  [ 9, 13]      - | 0.989564136601985
#>   [10]     seq1  [10, 14]      + | 0.130288870073855
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# grouping with other verbs
rng_by_strand %>% summarise(gc = mean(gc))

#> DataFrame with 3 rows and 2 columns
#>     strand        gc
#>   <factor> <numeric>
#> 1        + 0.5098712
#> 2        - 0.6304591
#> 3        * 0.6731152

rng_by_strand %>% filter(gc == min(gc))

#> GRanges object with 3 ranges and 1 metadata column:
#> Groups: 
#>       seqnames    ranges strand |                gc
#>          <Rle> <IRanges>  <Rle> |         <numeric>
#>   [1]     seq1  [10, 14]      + | 0.130288870073855
#>   [2]     seq1  [ 3,  7]      - | 0.348691981751472
#>   [3]     seq1  [ 7, 11]      * | 0.180407245177776
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# remove grouping with ungroup
rng_by_strand %>%
  ungroup() %>%
  summarise(gc = mean(gc))

#> DataFrame with 1 row and 1 column
#>          gc
#>   <numeric>
#> 1  0.623404