Read a BAM file

Read a BAM file

read_bam(file, index = file, paired = FALSE)

Arguments

file
A connection or path to a BAM file

index
The path to the BAM index file

paired
Whether to treat the BAM file as paired end (TRUE) or single end (FALSE).

Details

Reading a BAM file is deferred until an action such as selecting tags or filtering by overlapping ranges or filtering by tags is performed on it. If paired is set to TRUE, when alignments are load, the GRanges has two additional columns called read_pair_id and read_pair_group corresponding to paired reads (by default paired = TRUE will only select reads that are proper pairs).

For select valid columns are the either the fields of the BAM file. Valid entries are qname (QNAME), flag (FLAG), rname (RNAME), strand, pos (POS), qwidth (width of query), mapq (MAPQ), cigar (CIGAR), mrnm (RNEXT), mpos (PNEXT), isize (TLEN), seq (SEQ), and qual (QUAL). Any two character tags in the BAM file are also valid.

For filter the following fields are valid is_paired Select either unpaired (FALSE) or paired (TRUE) reads. is_proper_pair Select either improperly paired (FALSE) or properly paired (TRUE) reads. This is dependent on the alignment software used. is_unmapped_query Select unmapped (TRUE) or mapped (FALSE) reads. has_unmapped_mate Select reads with mapped (FALSE) or unmapped (TRUE) mates. is_minus_strand Select reads aligned to plus (FALSE) or minus (TRUE) strand. is_mate_minus_strand Select reads where mate is aligned to plus (FALSE) or minus (TRUE) strand. is_first_mate_read Select reads if they are the first mate (TRUE) or not (FALSE). is_second_mate_read Select reads if they are the second mate (TRUE) or not (FALSE). is_secondary_alignment Select reads if their alignment status is secondary (TRUE) or not (FALSE). This might be relevant if there are multimapping reads. is_not_passing_quality_controls Select reads that either pass quality controls (FALSE) or that do not (TRUE). is_duplicate Select reads that are unduplicated (FALSE) or duplicated (TRUE). This may represent reads that are PCR or optical duplicates.

Examples

if (require(pasillaBamSubset)) {
   bamfile <- untreated1_chr4()
   # nothing is read until an action has been performed
   print(read_bam(bamfile))
   # define a region of interest
   roi <- data.frame(seqnames = "chr4", start = 5e5, end = 7e5) %>%
            as_granges()
   # add map quality scores
   rng_mapq <- read_bam(bamfile) %>% select(mapq)
   print(rng_mapq)
   # filter_by_ovleraps will only read alignments if they overlap roi
   by_olap <- read_bam(bamfile) %>% filter_by_overlaps(roi)
   print(by_olap)
}

#> Loading required package: pasillaBamSubset

#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
#> GRanges object with 204355 ranges and 4 metadata columns:
#>            seqnames             ranges strand |      mapq       cigar    qwidth
#>               <Rle>          <IRanges>  <Rle> | <integer> <character> <integer>
#>        [1]     chr4        [892,  966]      - |       255         75M        75
#>        [2]     chr4        [919,  993]      - |       255         75M        75
#>        [3]     chr4        [924,  998]      + |         3         75M        75
#>        [4]     chr4        [936, 1010]      + |         3         75M        75
#>        [5]     chr4        [949, 1023]      + |         1         75M        75
#>        ...      ...                ...    ... .       ...         ...       ...
#>   [204351]     chr4 [1348268, 1348342]      + |         3         75M        75
#>   [204352]     chr4 [1348268, 1348342]      + |         3         75M        75
#>   [204353]     chr4 [1348268, 1348342]      + |         3         75M        75
#>   [204354]     chr4 [1348449, 1348523]      - |         0         75M        75
#>   [204355]     chr4 [1350124, 1350198]      - |         0         75M        75
#>                njunc
#>            <integer>
#>        [1]         0
#>        [2]         0
#>        [3]         0
#>        [4]         0
#>        [5]         0
#>        ...       ...
#>   [204351]         0
#>   [204352]         0
#>   [204353]         0
#>   [204354]         0
#>   [204355]         0
#>   -------
#>   seqinfo: 8 sequences from an unspecified genome
#> GRanges object with 28368 ranges and 4 metadata columns:
#>           seqnames           ranges strand |        which_label       cigar
#>              <Rle>        <IRanges>  <Rle> |              <Rle> <character>
#>       [1]     chr4 [499946, 500020]      + | chr4:500000-700000         75M
#>       [2]     chr4 [500023, 500097]      + | chr4:500000-700000         75M
#>       [3]     chr4 [500421, 500495]      - | chr4:500000-700000         75M
#>       [4]     chr4 [500684, 500758]      - | chr4:500000-700000         75M
#>       [5]     chr4 [500753, 500827]      - | chr4:500000-700000         75M
#>       ...      ...              ...    ... .                ...         ...
#>   [28364]     chr4 [699956, 700030]      + | chr4:500000-700000         75M
#>   [28365]     chr4 [699957, 700031]      + | chr4:500000-700000         75M
#>   [28366]     chr4 [699960, 700034]      - | chr4:500000-700000         75M
#>   [28367]     chr4 [699962, 700036]      + | chr4:500000-700000         75M
#>   [28368]     chr4 [699963, 700037]      + | chr4:500000-700000         75M
#>              qwidth     njunc
#>           <integer> <integer>
#>       [1]        75         0
#>       [2]        75         0
#>       [3]        75         0
#>       [4]        75         0
#>       [5]        75         0
#>       ...       ...       ...
#>   [28364]        75         0
#>   [28365]        75         0
#>   [28366]        75         0
#>   [28367]        75         0
#>   [28368]        75         0
#>   -------
#>   seqinfo: 8 sequences from an unspecified genome