Modify a Ranges object

Modify a Ranges object

# S3 method for Ranges
mutate(.data, ...)

Arguments

.data
a Ranges object

...
Pairs of name-value expressions. The name-value pairs can either create new metadata columns or modify existing ones.

Value

a Ranges object

Examples

df <- data.frame(start = 1:10,
                 width = 5,
                 seqnames = "seq1",
                 strand = sample(c("+", "-", "*"), 10, replace = TRUE),
                 gc = runif(10))
rng <- as_granges(df)

# mutate adds new columns
rng %>%
    mutate(avg_gc = mean(gc), row_id = 1:n())

#> GRanges object with 10 ranges and 3 metadata columns:
#>        seqnames    ranges strand |                gc            avg_gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>         <numeric>
#>    [1]     seq1       1-5      - | 0.330660525709391 0.607461047288962
#>    [2]     seq1       2-6      * | 0.865120546659455 0.607461047288962
#>    [3]     seq1       3-7      - | 0.777584439376369 0.607461047288962
#>    [4]     seq1       4-8      * | 0.827303449623287 0.607461047288962
#>    [5]     seq1       5-9      * | 0.603324356488883 0.607461047288962
#>    [6]     seq1      6-10      * | 0.491231821943074 0.607461047288962
#>    [7]     seq1      7-11      + | 0.780358511023223 0.607461047288962
#>    [8]     seq1      8-12      - | 0.884227027418092 0.607461047288962
#>    [9]     seq1      9-13      * | 0.207713897805661 0.607461047288962
#>   [10]     seq1     10-14      + | 0.307085896842182 0.607461047288962
#>           row_id
#>        <integer>
#>    [1]         1
#>    [2]         2
#>    [3]         3
#>    [4]         4
#>    [5]         5
#>    [6]         6
#>    [7]         7
#>    [8]         8
#>    [9]         9
#>   [10]        10
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# can also compute on newly created columns
rng %>%
    mutate(score = gc * width, score2 = score + 1)

#> GRanges object with 10 ranges and 3 metadata columns:
#>        seqnames    ranges strand |                gc            score
#>           <Rle> <IRanges>  <Rle> |         <numeric>        <numeric>
#>    [1]     seq1       1-5      - | 0.330660525709391 1.65330262854695
#>    [2]     seq1       2-6      * | 0.865120546659455 4.32560273329727
#>    [3]     seq1       3-7      - | 0.777584439376369 3.88792219688185
#>    [4]     seq1       4-8      * | 0.827303449623287 4.13651724811643
#>    [5]     seq1       5-9      * | 0.603324356488883 3.01662178244442
#>    [6]     seq1      6-10      * | 0.491231821943074 2.45615910971537
#>    [7]     seq1      7-11      + | 0.780358511023223 3.90179255511612
#>    [8]     seq1      8-12      - | 0.884227027418092 4.42113513709046
#>    [9]     seq1      9-13      * | 0.207713897805661  1.0385694890283
#>   [10]     seq1     10-14      + | 0.307085896842182 1.53542948421091
#>                  score2
#>               <numeric>
#>    [1] 2.65330262854695
#>    [2] 5.32560273329727
#>    [3] 4.88792219688185
#>    [4] 5.13651724811643
#>    [5] 4.01662178244442
#>    [6] 3.45615910971537
#>    [7] 4.90179255511612
#>    [8] 5.42113513709046
#>    [9]  2.0385694890283
#>   [10] 2.53542948421091
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# group by partitions the data and computes within each group
rng %>%
    group_by(strand) %>%
    mutate(avg_gc = mean(gc), row_id = 1:n())

#> GRanges object with 10 ranges and 3 metadata columns:
#> Groups: strand [3]
#>        seqnames    ranges strand |                gc            avg_gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>         <numeric>
#>    [1]     seq1       1-5      - | 0.330660525709391 0.664157330834617
#>    [2]     seq1       2-6      * | 0.865120546659455 0.598938814504072
#>    [3]     seq1       3-7      - | 0.777584439376369 0.664157330834617
#>    [4]     seq1       4-8      * | 0.827303449623287 0.598938814504072
#>    [5]     seq1       5-9      * | 0.603324356488883 0.598938814504072
#>    [6]     seq1      6-10      * | 0.491231821943074 0.598938814504072
#>    [7]     seq1      7-11      + | 0.780358511023223 0.543722203932703
#>    [8]     seq1      8-12      - | 0.884227027418092 0.664157330834617
#>    [9]     seq1      9-13      * | 0.207713897805661 0.598938814504072
#>   [10]     seq1     10-14      + | 0.307085896842182 0.543722203932703
#>           row_id
#>        <integer>
#>    [1]         1
#>    [2]         1
#>    [3]         2
#>    [4]         2
#>    [5]         3
#>    [6]         4
#>    [7]         1
#>    [8]         3
#>    [9]         5
#>   [10]         2
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# mutate can be used in conjuction with anchoring to resize ranges
rng %>%
    mutate(width = 10)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      1-10      - | 0.330660525709391
#>    [2]     seq1      2-11      * | 0.865120546659455
#>    [3]     seq1      3-12      - | 0.777584439376369
#>    [4]     seq1      4-13      * | 0.827303449623287
#>    [5]     seq1      5-14      * | 0.603324356488883
#>    [6]     seq1      6-15      * | 0.491231821943074
#>    [7]     seq1      7-16      + | 0.780358511023223
#>    [8]     seq1      8-17      - | 0.884227027418092
#>    [9]     seq1      9-18      * | 0.207713897805661
#>   [10]     seq1     10-19      + | 0.307085896842182
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# by default width modfication fixes by start
rng %>%
    anchor_start() %>%
    mutate(width = 10)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      1-10      - | 0.330660525709391
#>    [2]     seq1      2-11      * | 0.865120546659455
#>    [3]     seq1      3-12      - | 0.777584439376369
#>    [4]     seq1      4-13      * | 0.827303449623287
#>    [5]     seq1      5-14      * | 0.603324356488883
#>    [6]     seq1      6-15      * | 0.491231821943074
#>    [7]     seq1      7-16      + | 0.780358511023223
#>    [8]     seq1      8-17      - | 0.884227027418092
#>    [9]     seq1      9-18      * | 0.207713897805661
#>   [10]     seq1     10-19      + | 0.307085896842182
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# fix by end or midpoint
rng %>%
    anchor_end() %>%
    mutate(width = width + 1)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1       0-5      - | 0.330660525709391
#>    [2]     seq1       1-6      * | 0.865120546659455
#>    [3]     seq1       2-7      - | 0.777584439376369
#>    [4]     seq1       3-8      * | 0.827303449623287
#>    [5]     seq1       4-9      * | 0.603324356488883
#>    [6]     seq1      5-10      * | 0.491231821943074
#>    [7]     seq1      6-11      + | 0.780358511023223
#>    [8]     seq1      7-12      - | 0.884227027418092
#>    [9]     seq1      8-13      * | 0.207713897805661
#>   [10]     seq1      9-14      + | 0.307085896842182
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

rng %>%
    anchor_center() %>%
    mutate(width = width + 1)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1       0-5      - | 0.330660525709391
#>    [2]     seq1       1-6      * | 0.865120546659455
#>    [3]     seq1       2-7      - | 0.777584439376369
#>    [4]     seq1       3-8      * | 0.827303449623287
#>    [5]     seq1       4-9      * | 0.603324356488883
#>    [6]     seq1      5-10      * | 0.491231821943074
#>    [7]     seq1      6-11      + | 0.780358511023223
#>    [8]     seq1      7-12      - | 0.884227027418092
#>    [9]     seq1      8-13      * | 0.207713897805661
#>   [10]     seq1      9-14      + | 0.307085896842182
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# anchoring by strand
rng %>%
    anchor_3p() %>%
    mutate(width = width * 2)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      1-10      - | 0.330660525709391
#>    [2]     seq1      -3-6      * | 0.865120546659455
#>    [3]     seq1      3-12      - | 0.777584439376369
#>    [4]     seq1      -1-8      * | 0.827303449623287
#>    [5]     seq1       0-9      * | 0.603324356488883
#>    [6]     seq1      1-10      * | 0.491231821943074
#>    [7]     seq1      2-11      + | 0.780358511023223
#>    [8]     seq1      8-17      - | 0.884227027418092
#>    [9]     seq1      4-13      * | 0.207713897805661
#>   [10]     seq1      5-14      + | 0.307085896842182
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

rng %>%
    anchor_5p() %>%
    mutate(width = width * 2)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      -4-5      - | 0.330660525709391
#>    [2]     seq1      2-11      * | 0.865120546659455
#>    [3]     seq1      -2-7      - | 0.777584439376369
#>    [4]     seq1      4-13      * | 0.827303449623287
#>    [5]     seq1      5-14      * | 0.603324356488883
#>    [6]     seq1      6-15      * | 0.491231821943074
#>    [7]     seq1      7-16      + | 0.780358511023223
#>    [8]     seq1      3-12      - | 0.884227027418092
#>    [9]     seq1      9-18      * | 0.207713897805661
#>   [10]     seq1     10-19      + | 0.307085896842182
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths