Modify a Ranges object

Modify a Ranges object

# S3 method for Ranges
mutate(.data, ...)

Arguments

.data
a Ranges object

...
Pairs of name-value expressions. The name-value pairs can either create new metadata columns or modify existing ones.

Value

a Ranges object

Examples

df <- data.frame(start = 1:10,
                 width = 5,
                 seqnames = "seq1",
                 strand = sample(c("+", "-", "*"), 10, replace = TRUE),
                 gc = runif(10))
rng <- as_granges(df)

# mutate adds new columns
rng %>%
    mutate(avg_gc = mean(gc), row_id = 1:n())

#> GRanges object with 10 ranges and 3 metadata columns:
#>        seqnames    ranges strand |                gc            avg_gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>         <numeric>
#>    [1]     seq1       1-5      - | 0.884227027418092 0.340701564052142
#>    [2]     seq1       2-6      * | 0.207713897805661 0.340701564052142
#>    [3]     seq1       3-7      - | 0.307085896842182 0.340701564052142
#>    [4]     seq1       4-8      + | 0.330529848113656 0.340701564052142
#>    [5]     seq1       5-9      * |  0.19867907022126 0.340701564052142
#>    [6]     seq1      6-10      * | 0.235694302013144 0.340701564052142
#>    [7]     seq1      7-11      + | 0.274886660277843 0.340701564052142
#>    [8]     seq1      8-12      * | 0.591321053681895 0.340701564052142
#>    [9]     seq1      9-13      - | 0.253390653757378 0.340701564052142
#>   [10]     seq1     10-14      - |  0.12348723039031 0.340701564052142
#>           row_id
#>        <integer>
#>    [1]         1
#>    [2]         2
#>    [3]         3
#>    [4]         4
#>    [5]         5
#>    [6]         6
#>    [7]         7
#>    [8]         8
#>    [9]         9
#>   [10]        10
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# can also compute on newly created columns
rng %>%
    mutate(score = gc * width, score2 = score + 1)

#> GRanges object with 10 ranges and 3 metadata columns:
#>        seqnames    ranges strand |                gc             score
#>           <Rle> <IRanges>  <Rle> |         <numeric>         <numeric>
#>    [1]     seq1       1-5      - | 0.884227027418092  4.42113513709046
#>    [2]     seq1       2-6      * | 0.207713897805661   1.0385694890283
#>    [3]     seq1       3-7      - | 0.307085896842182  1.53542948421091
#>    [4]     seq1       4-8      + | 0.330529848113656  1.65264924056828
#>    [5]     seq1       5-9      * |  0.19867907022126 0.993395351106301
#>    [6]     seq1      6-10      * | 0.235694302013144  1.17847151006572
#>    [7]     seq1      7-11      + | 0.274886660277843  1.37443330138922
#>    [8]     seq1      8-12      * | 0.591321053681895  2.95660526840948
#>    [9]     seq1      9-13      - | 0.253390653757378  1.26695326878689
#>   [10]     seq1     10-14      - |  0.12348723039031 0.617436151951551
#>                  score2
#>               <numeric>
#>    [1] 5.42113513709046
#>    [2]  2.0385694890283
#>    [3] 2.53542948421091
#>    [4] 2.65264924056828
#>    [5]  1.9933953511063
#>    [6] 2.17847151006572
#>    [7] 2.37443330138922
#>    [8] 3.95660526840948
#>    [9] 2.26695326878689
#>   [10] 1.61743615195155
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# group by partitions the data and computes within each group
rng %>%
    group_by(strand) %>%
    mutate(avg_gc = mean(gc), row_id = 1:n())

#> GRanges object with 10 ranges and 3 metadata columns:
#> Groups: strand [3]
#>        seqnames    ranges strand |                gc            avg_gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>         <numeric>
#>    [1]     seq1       1-5      - | 0.884227027418092 0.392047702101991
#>    [2]     seq1       2-6      * | 0.207713897805661  0.30835208093049
#>    [3]     seq1       3-7      - | 0.307085896842182 0.392047702101991
#>    [4]     seq1       4-8      + | 0.330529848113656  0.30270825419575
#>    [5]     seq1       5-9      * |  0.19867907022126  0.30835208093049
#>    [6]     seq1      6-10      * | 0.235694302013144  0.30835208093049
#>    [7]     seq1      7-11      + | 0.274886660277843  0.30270825419575
#>    [8]     seq1      8-12      * | 0.591321053681895  0.30835208093049
#>    [9]     seq1      9-13      - | 0.253390653757378 0.392047702101991
#>   [10]     seq1     10-14      - |  0.12348723039031 0.392047702101991
#>           row_id
#>        <integer>
#>    [1]         1
#>    [2]         1
#>    [3]         2
#>    [4]         1
#>    [5]         2
#>    [6]         3
#>    [7]         2
#>    [8]         4
#>    [9]         3
#>   [10]         4
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# mutate can be used in conjuction with anchoring to resize ranges
rng %>%
    mutate(width = 10)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      1-10      - | 0.884227027418092
#>    [2]     seq1      2-11      * | 0.207713897805661
#>    [3]     seq1      3-12      - | 0.307085896842182
#>    [4]     seq1      4-13      + | 0.330529848113656
#>    [5]     seq1      5-14      * |  0.19867907022126
#>    [6]     seq1      6-15      * | 0.235694302013144
#>    [7]     seq1      7-16      + | 0.274886660277843
#>    [8]     seq1      8-17      * | 0.591321053681895
#>    [9]     seq1      9-18      - | 0.253390653757378
#>   [10]     seq1     10-19      - |  0.12348723039031
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# by default width modfication fixes by start
rng %>%
    anchor_start() %>%
    mutate(width = 10)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      1-10      - | 0.884227027418092
#>    [2]     seq1      2-11      * | 0.207713897805661
#>    [3]     seq1      3-12      - | 0.307085896842182
#>    [4]     seq1      4-13      + | 0.330529848113656
#>    [5]     seq1      5-14      * |  0.19867907022126
#>    [6]     seq1      6-15      * | 0.235694302013144
#>    [7]     seq1      7-16      + | 0.274886660277843
#>    [8]     seq1      8-17      * | 0.591321053681895
#>    [9]     seq1      9-18      - | 0.253390653757378
#>   [10]     seq1     10-19      - |  0.12348723039031
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# fix by end or midpoint
rng %>%
    anchor_end() %>%
    mutate(width = width + 1)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1       0-5      - | 0.884227027418092
#>    [2]     seq1       1-6      * | 0.207713897805661
#>    [3]     seq1       2-7      - | 0.307085896842182
#>    [4]     seq1       3-8      + | 0.330529848113656
#>    [5]     seq1       4-9      * |  0.19867907022126
#>    [6]     seq1      5-10      * | 0.235694302013144
#>    [7]     seq1      6-11      + | 0.274886660277843
#>    [8]     seq1      7-12      * | 0.591321053681895
#>    [9]     seq1      8-13      - | 0.253390653757378
#>   [10]     seq1      9-14      - |  0.12348723039031
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

rng %>%
    anchor_center() %>%
    mutate(width = width + 1)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1       0-5      - | 0.884227027418092
#>    [2]     seq1       1-6      * | 0.207713897805661
#>    [3]     seq1       2-7      - | 0.307085896842182
#>    [4]     seq1       3-8      + | 0.330529848113656
#>    [5]     seq1       4-9      * |  0.19867907022126
#>    [6]     seq1      5-10      * | 0.235694302013144
#>    [7]     seq1      6-11      + | 0.274886660277843
#>    [8]     seq1      7-12      * | 0.591321053681895
#>    [9]     seq1      8-13      - | 0.253390653757378
#>   [10]     seq1      9-14      - |  0.12348723039031
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# anchoring by strand
rng %>%
    anchor_3p() %>%
    mutate(width = width * 2)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      1-10      - | 0.884227027418092
#>    [2]     seq1      -3-6      * | 0.207713897805661
#>    [3]     seq1      3-12      - | 0.307085896842182
#>    [4]     seq1      -1-8      + | 0.330529848113656
#>    [5]     seq1       0-9      * |  0.19867907022126
#>    [6]     seq1      1-10      * | 0.235694302013144
#>    [7]     seq1      2-11      + | 0.274886660277843
#>    [8]     seq1      3-12      * | 0.591321053681895
#>    [9]     seq1      9-18      - | 0.253390653757378
#>   [10]     seq1     10-19      - |  0.12348723039031
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

rng %>%
    anchor_5p() %>%
    mutate(width = width * 2)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                gc
#>           <Rle> <IRanges>  <Rle> |         <numeric>
#>    [1]     seq1      -4-5      - | 0.884227027418092
#>    [2]     seq1      2-11      * | 0.207713897805661
#>    [3]     seq1      -2-7      - | 0.307085896842182
#>    [4]     seq1      4-13      + | 0.330529848113656
#>    [5]     seq1      5-14      * |  0.19867907022126
#>    [6]     seq1      6-15      * | 0.235694302013144
#>    [7]     seq1      7-16      + | 0.274886660277843
#>    [8]     seq1      8-17      * | 0.591321053681895
#>    [9]     seq1      4-13      - | 0.253390653757378
#>   [10]     seq1      5-14      - |  0.12348723039031
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths