Find nearest neighbours between two Ranges objects

Find nearest neighbours between two Ranges objects

join_nearest(x, y, suffix = c(".x", ".y"))

join_nearest_left(x, y, suffix = c(".x", ".y"))

join_nearest_right(x, y, suffix = c(".x", ".y"))

join_nearest_upstream(x, y, suffix = c(".x", ".y"))

join_nearest_downstream(x, y, suffix = c(".x", ".y"))

Arguments

x, y
Ranges objects, add the nearest neighbours of ranges in x to those in y.

suffix
A character vector of length two used to identify metadata columns coming from x and y.

Value

A Ranges object corresponding to the nearest ranges, all metadata is copied over from the right-hand side ranges `y`.

Details

By default join_nearest will find abritrary nearest neighbours in either direction and ignore any strand information. The join_nearest_left and join_nearest_right methods will find abritrary nearest neighbour ranges on x that are left/right of those on y and ignore any strand information.

The join_nearest_upstream method will find abritrary nearest neighbour ranges on x that are upstream of those on y. This takes into account strandedness of the ranges. On the positive strand nearest upstream will be on the left and on the negative strand nearest upstream will be on the right.

The join_nearest_downstream method will find abritrary nearest neighbour ranges on x that are upstream of those on y. This takes into account strandedness of the ranges.On the positive strand nearest downstream will be on the right and on the negative strand nearest upstream will be on the left.