Select metadata columns of the Ranges object by name or position

Select metadata columns of the Ranges object by name or position

# S3 method for GenomicRanges
select(.data, ..., .drop_ranges = FALSE)

# S3 method for Ranges
select(.data, ..., .drop_ranges = FALSE)

Arguments

.data
a Ranges object

...
One or more metadata column names.

.drop_ranges
If TRUE select will always return a tibble. In this case, you may select columns that form the core part of the Ranges object.

Value

a Ranges object or a tibble

Details

Note that by default select only acts on the metadata columns (and will therefore return a Ranges object) if a core component of a Ranges is dropped or selected without the other required components (this includes the seqnames, strand, start, end, width names), then select will throw an error unless .drop_ranges is set to TRUE.

See also

select

Examples

df <- data.frame(start = 1:10, width = 5,  seqnames = "seq1",
strand = sample(c("+", "-", "*"), 10, replace = TRUE), gc = runif(10), counts = rpois(10, 2))
rng <- Ranges(df)
select(rng, -gc)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |    counts
#>           <Rle> <IRanges>  <Rle> | <integer>
#>    [1]     seq1  [ 1,  5]      - |         1
#>    [2]     seq1  [ 2,  6]      - |         1
#>    [3]     seq1  [ 3,  7]      * |         0
#>    [4]     seq1  [ 4,  8]      * |         1
#>    [5]     seq1  [ 5,  9]      + |         2
#>    [6]     seq1  [ 6, 10]      - |         3
#>    [7]     seq1  [ 7, 11]      * |         5
#>    [8]     seq1  [ 8, 12]      + |         3
#>    [9]     seq1  [ 9, 13]      + |         0
#>   [10]     seq1  [10, 14]      * |         2
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

select(rng, gc)

#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |                 gc
#>           <Rle> <IRanges>  <Rle> |          <numeric>
#>    [1]     seq1  [ 1,  5]      - |  0.906954375561327
#>    [2]     seq1  [ 2,  6]      - |  0.209816768066958
#>    [3]     seq1  [ 3,  7]      * |  0.358137989183888
#>    [4]     seq1  [ 4,  8]      * |  0.448299144394696
#>    [5]     seq1  [ 5,  9]      + |  0.906426433008164
#>    [6]     seq1  [ 6, 10]      - |  0.389439295744523
#>    [7]     seq1  [ 7, 11]      * |  0.517459749476984
#>    [8]     seq1  [ 8, 12]      + |  0.125239087734371
#>    [9]     seq1  [ 9, 13]      + | 0.0301457457244396
#>   [10]     seq1  [10, 14]      * |  0.771805494558066
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

select(rng, counts, gc)

#> GRanges object with 10 ranges and 2 metadata columns:
#>        seqnames    ranges strand |    counts                 gc
#>           <Rle> <IRanges>  <Rle> | <integer>          <numeric>
#>    [1]     seq1  [ 1,  5]      - |         1  0.906954375561327
#>    [2]     seq1  [ 2,  6]      - |         1  0.209816768066958
#>    [3]     seq1  [ 3,  7]      * |         0  0.358137989183888
#>    [4]     seq1  [ 4,  8]      * |         1  0.448299144394696
#>    [5]     seq1  [ 5,  9]      + |         2  0.906426433008164
#>    [6]     seq1  [ 6, 10]      - |         3  0.389439295744523
#>    [7]     seq1  [ 7, 11]      * |         5  0.517459749476984
#>    [8]     seq1  [ 8, 12]      + |         3  0.125239087734371
#>    [9]     seq1  [ 9, 13]      + |         0 0.0301457457244396
#>   [10]     seq1  [10, 14]      * |         2  0.771805494558066
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

select(rng, 2:1)

#> GRanges object with 10 ranges and 2 metadata columns:
#>        seqnames    ranges strand |    counts                 gc
#>           <Rle> <IRanges>  <Rle> | <integer>          <numeric>
#>    [1]     seq1  [ 1,  5]      - |         1  0.906954375561327
#>    [2]     seq1  [ 2,  6]      - |         1  0.209816768066958
#>    [3]     seq1  [ 3,  7]      * |         0  0.358137989183888
#>    [4]     seq1  [ 4,  8]      * |         1  0.448299144394696
#>    [5]     seq1  [ 5,  9]      + |         2  0.906426433008164
#>    [6]     seq1  [ 6, 10]      - |         3  0.389439295744523
#>    [7]     seq1  [ 7, 11]      * |         5  0.517459749476984
#>    [8]     seq1  [ 8, 12]      + |         3  0.125239087734371
#>    [9]     seq1  [ 9, 13]      + |         0 0.0301457457244396
#>   [10]     seq1  [10, 14]      * |         2  0.771805494558066
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

select(rng, seqnames, strand, .drop_ranges = TRUE)

#> # A tibble: 10 x 2
#>    seqnames strand
#>      <fctr> <fctr>
#>  1     seq1      -
#>  2     seq1      -
#>  3     seq1      *
#>  4     seq1      *
#>  5     seq1      +
#>  6     seq1      -
#>  7     seq1      *
#>  8     seq1      +
#>  9     seq1      +
#> 10     seq1      *